+++ lib/strtoul.c
@@ -45,7 +45,7 @@ unsigned long lib_strtoul(char *nptr, char **endptr, int base)
 	}
 
 	while (strtoul_isalnum(*nptr) != 0) {
-		t = (unsigned int)((unsigned int)*nptr - (unsigned int)'0');
+		t = (unsigned int)*nptr - (unsigned int)'0';
 		if (t > 9U) {
 			t = ((unsigned int)*nptr | 0x20U) - (unsigned int)'a' + 10U;
 		}
